folder structure refactor
- added a new `withdrawal` feature folder with the appropriate structure
