+++ lib/apps/artisan_app/features/withdrawal/ui/widgets/wallet_balance_card.dart
